Adpiot V2.1 demo/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.14.0
python-Levenshtein==0.25.1
langchain==0.1.10
chromadb==0.4.24
langgraph==0.1.0


